defense1/requirements.txt
absl-py==2.3.1
adversarial-robustness-toolbox==1.20.1
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
git-filter-repo==2.47.0
grpcio==1.76.0
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
llvmlite==0.46.0
Markdown==3.10
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
nvitop==1.6.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
psutil==7.2.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.12.20
torch==2.9.1+cu126
torchvision==0.24.1+cu126
tqdm==4.67.1
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.3
vit-pytorch==1.17.1
Werkzeug==3.1.4
